Refactor - moved ZuberasDiedWatcher to shared area
