+++ Mage.Sets/src/mage/sets/championsofkamigawa/FloatingDreamZubera.java
@@ -29,7 +29,6 @@
 package mage.sets.championsofkamigawa;
 
 import mage.MageInt;
-import mage.MageObject;
 import mage.abilities.Ability;
 import mage.abilities.common.DiesTriggeredAbility;
 import mage.abilities.dynamicvalue.DynamicValue;
@@ -38,12 +37,8 @@ import mage.abilities.effects.common.DrawCardSourceControllerEffect;
 import mage.cards.CardImpl;
 import mage.constants.CardType;
 import mage.constants.Rarity;
-import mage.constants.WatcherScope;
-import mage.constants.Zone;
 import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.events.ZoneChangeEvent;
-import mage.watchers.Watcher;
+import mage.watchers.common.ZuberasDiedWatcher;
 
 import java.util.UUID;
 
@@ -74,42 +69,6 @@ public class FloatingDreamZubera extends CardImpl {
 
 }
 
-
-class ZuberasDiedWatcher extends Watcher {
-
-    public int zuberasDiedThisTurn = 0;
-
-    public ZuberasDiedWatcher() {
-        super("ZuberasDied", WatcherScope.GAME);
-    }
-
-    public ZuberasDiedWatcher(final ZuberasDiedWatcher watcher) {
-        super(watcher);
-        this.zuberasDiedThisTurn = watcher.zuberasDiedThisTurn;
-    }
-
-    @Override
-    public ZuberasDiedWatcher copy() {
-        return new ZuberasDiedWatcher(this);
-    }
-
-    @Override
-    public void watch(GameEvent event, Game game) {
-        if (event.getType() == GameEvent.EventType.ZONE_CHANGE && ((ZoneChangeEvent) event).isDiesEvent()) {
-            MageObject card = game.getLastKnownInformation(event.getTargetId(), Zone.BATTLEFIELD);
-            if (card != null && card.hasSubtype("Zubera", game)) {
-                zuberasDiedThisTurn++;
-            }
-        }
-    }
-
-    @Override
-    public void reset() {
-        super.reset();
-        zuberasDiedThisTurn = 0;
-    }
-
-}
 
 class ZuberasDiedDynamicValue implements DynamicValue {
